divide map logic into smart and dumb map
