+++ .gitignore
@@ -23,4 +23,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-yarn.lock
+yarn.lock.json
+package-lock.json
