+++ src/App.scss
@@ -18,6 +18,9 @@ body{
 }
 
 .App{
+  background-image: url('./city.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
   height: 100%;
   width: 100%;
   display: flex;
@@ -27,53 +30,11 @@ body{
     margin: 0;
     padding: 0;
   }
-  & > .container{
+  & > .main{
     flex:1;
-  }
-
-  .tooltip-content{
-    direction: rtl;
-    .images{
-      width: 100%;
-      display: flex;
-      justify-content: space-between;
-      font-size: 2rem;
-      color: dodgerblue;
-      .icon-bounce {
-        cursor: pointer;
-        animation: bounce .5s infinite alternate;
-        -webkit-animation: bounce .5s infinite alternate;
-        transform: translateY(4px);
-      }
-
-      .icon-spin{
-        cursor: pointer;
-        transform:rotate(90deg);
-      }
-    }
-    .description{
-      width: 100%;
-      .key{
-        font-weight: bold;
-      }
-    }
-  }
-
-  @keyframes bounce {
-    from {
-      transform: translateY(4px);
-    }
-    to {
-      transform: translateY(0px);
-    }
-  }
-  @-webkit-keyframes bounce {
-    from {
-      transform: translateY(4px);
-    }
-    to {
-      transform: translateY(0px);
-    }
+    display: flex;
+    align-items: center;
+    justify-content: center;
   }
 }
