+++ src/index.js
@@ -7,9 +7,9 @@ import * as serviceWorker from './serviceWorker';
 const root = document.getElementById('root');
 if(root){
     ReactDOM.render(<App />, root);
+    // If you want your app to work offline and load faster, you can change
+    // unregister() to register() below. Note this comes with some pitfalls.
+    // Learn more about service workers: http://bit.ly/CRA-PWA
+    serviceWorker.unregister();
 }
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
