--- src/map/Map.css
@@ -1,5 +1,0 @@
-.map-container,
-.map-instance{
-    height: 300px;
-    width: 300px;
-}
